Don't use ptrdiff_t in tests of maxElement, etc.

+++ branches/v0.70/test/TMV_TestSmallVector.cpp
@@ -51,7 +51,7 @@ static void TestSmallVectorReal()
     v(23) = T(10*N);
     v(42) = T(1)/T(4);
     v(15) = T(-20*N);
-    ptrdiff_t imax,imin;
+    int imax,imin;
     Assert(v.maxAbsElement(&imax) == T(20*N),
            "MaxAbsElement of SmallVector did not return correct value");
     Assert(imax == 15,
@@ -241,7 +241,7 @@ static void TestSmallVectorComplex()
     v3(23) = CT(40*N,9*N);
     v3(42) = CT(0,1);
     v3(15) = CT(-32*N,24*N);
-    ptrdiff_t imax,imin;
+    int imax,imin;
     if (showacc) {
         std::cout<<"v = "<<v3<<std::endl;
         std::cout<<"v.MaxAbs = "<<v3.maxAbsElement(&imax)<<std::endl;
